requirements.txt
python-dotenv
requests
tradingview-ta
pydantic
aiogram
openai
yfinance
pandas
numpy
scikit-learn
joblib
